+++ tests/reduction.rs
@@ -6,6 +6,7 @@ use image::Pixel;
 use oxipng::colors::{BitDepth, ColorType};
 use oxipng::png;
 use std::collections::HashSet;
+use std::error::Error;
 use std::fs::remove_file;
 use std::path::Path;
 use std::path::PathBuf;
@@ -37,7 +38,7 @@ fn test_it_converts(input: &Path,
 
     match oxipng::optimize(input, opts) {
         Ok(_) => (),
-        Err(x) => panic!(x.to_owned()),
+        Err(x) => panic!(x.description().to_owned()),
     };
     assert!(output.exists());
 
@@ -45,7 +46,7 @@ fn test_it_converts(input: &Path,
         Ok(x) => x,
         Err(x) => {
             remove_file(output).ok();
-            panic!(x.to_owned())
+            panic!(x.description().to_owned())
         }
     };
 
@@ -1247,7 +1248,7 @@ fn palette_should_be_reduced_with_dupes() {
 
     match oxipng::optimize(&input, &opts) {
         Ok(_) => (),
-        Err(x) => panic!(x.to_owned()),
+        Err(x) => panic!(x.description().to_owned()),
     };
     assert!(output.exists());
 
@@ -1255,7 +1256,7 @@ fn palette_should_be_reduced_with_dupes() {
         Ok(x) => x,
         Err(x) => {
             remove_file(&output).ok();
-            panic!(x.to_owned())
+            panic!(x.description().to_owned())
         }
     };
 
@@ -1287,7 +1288,7 @@ fn palette_should_be_reduced_with_unused() {
 
     match oxipng::optimize(&input, &opts) {
         Ok(_) => (),
-        Err(x) => panic!(x.to_owned()),
+        Err(x) => panic!(x.description().to_owned()),
     };
     assert!(output.exists());
 
@@ -1295,7 +1296,7 @@ fn palette_should_be_reduced_with_unused() {
         Ok(x) => x,
         Err(x) => {
             remove_file(&output).ok();
-            panic!(x.to_owned())
+            panic!(x.description().to_owned())
         }
     };
 
@@ -1327,7 +1328,7 @@ fn palette_should_be_reduced_with_both() {
 
     match oxipng::optimize(&input, &opts) {
         Ok(_) => (),
-        Err(x) => panic!(x.to_owned()),
+        Err(x) => panic!(x.description().to_owned()),
     };
     assert!(output.exists());
 
@@ -1335,7 +1336,7 @@ fn palette_should_be_reduced_with_both() {
         Ok(x) => x,
         Err(x) => {
             remove_file(&output).ok();
-            panic!(x.to_owned())
+            panic!(x.description().to_owned())
         }
     };