+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 **Version 0.11.0 (unreleased)**
- - [SEMVER_MAJOR] Bump minimum rustc version to 1.8.0, required by dependencies
+ - [SEMVER_MAJOR] Bump minimum rustc version to 1.9.0, required by dependencies
  - [SEMVER_MINOR] Allow calling optimization presets via crate using `Options::from_preset`
+ - [SEMVER_MAJOR] Return proper `PngError` type which implements `std::error::Error` from `Result`s
 
 **Version 0.10.0**
  - [SEMVER_MINOR] Make clap and regex dependencies optional